Migrate node attributes and extras to JSONB for Django backend
With Django adding support for the Posgres JSONB field, the old custom
implementation of the entity-attribute-value (EAV) model, used for the
attributes and extras for the `DbNode` table, has been made obsolete.

Here we change the `DbNode` model for the Django backend to use the
`JSONB` type directly and migrations are added to change the schema and
migrate the existing data.

The `DbSetting` table also used the custom EAV model and has likewise
been migrated to use a JSONB field instead.

The hard coded dummy model that the query builder uses to map the Django
database models onto SqlAlchemy variants, has been removed and replaced
by Aldjemy models that are generated on the fly.